unsigned is 32-bits. so use use uin64_t where necessary

+++ mee_decryptor.h
@@ -31,7 +31,7 @@ const unsigned REQUESTS_PER_BLOCK = 8;
 class Decryptor: public SimObject{
     
 public:
-    Decryptor(FACache *cache_, MEESystem* dram_, unsigned, unsigned);
+    Decryptor(FACache *cache_, MEESystem* dram_, uint64_t, uint64_t);
 
     bool can_accept_input();
     bool is_output_ready();
